requirements (9).txt
streamlit>=1.32.0
supabase
pandas
plotly
Pillow
requests
gTTS
google-generativeai
